ROM image file loader/manager from SD card. Emulator reset button, refactoring, preparation for bare-metal, and bug fixes.

+++ mem.c
@@ -181,7 +181,7 @@ int  mem_define_io(int addr_start, int addr_end, io_handler_callback io_handler)
  *  return: ' 0' - write ok,
  *          '-1' - memory location is out of range
  */
-int mem_load(int addr_start, int *buffer, int length)
+int mem_load(int addr_start, uint8_t *buffer, int length)
 {
     int i;
 
@@ -191,7 +191,7 @@ int mem_load(int addr_start, int *buffer, int length)
 
     for (i = 0; i < length; i++)
     {
-        memory[(i+addr_start)].data_byte = (uint8_t)buffer[i];
+        memory[(i+addr_start)].data_byte = buffer[i];
     }
 
     return MEM_OK;
